# [0.4.0]: 22-04-2026

[Added]
- Finished cleaning up the novami.deep module; all classes, functions, and DL templates *should* work well with each other
- Partially missing documentation in the novami.deep.vectorizer and utils
- Unit test suite for molecular descriptors, distance calculations, partition, and data loading of deep module

[Fixed]
- Fixed the swapped GetFingerprintAsNumpy and GetCountFingerprintAsNumpy calls in smiles_2_* when a single SMILES string was passed
- Non-compatible Float32 assignment in outlier filter. Now uses Float64

[Removed]
- The remaining old code from the novami.deep module

# [0.3.5]: 01-04-2026

[Added]
- Extend functionality of group_k_neighbors_distance
- Add W.I.P. Winsorization to DataTransformer
- (Partially) clean the deep module
- ``MMDataset`` / ``MMBatch``, ``MMLoader``, and ``MMTUnit`` as the primary deep stack; legacy ``GNNRegressor`` and ``MMGV`` moved under ``deprecated/deep/`` with optional re-exports
- Calculations for Precision-Recall Curve AUC, Expected Calibration Error, Brier Score
- New QCG template for KFold-like calculations

[Fixed]
- The MAD deduplication procedure is now used

[Removed]
- A lot of old code from the novami.deep module
- ``novami/deep/model.py`` (MMMTGNN and backbones); use ``deprecated/deep/mmmtgnn.py`` instead

+++ novami/data/filter.py
@@ -48,10 +48,10 @@ def filter_outliers(df: pl.DataFrame, smiles_col: str = 'SMILES', kwargs: Dict =
     )
 
     df = df.with_columns([
-        pl.col('Mol').map_elements(Descriptors.MolWt, return_dtype=pl.Float32).alias('MolWt'),
-        pl.col('Mol').map_elements(Descriptors.MolLogP, return_dtype=pl.Float32).alias('MolLogP'),
-        pl.col('Mol').map_elements(Descriptors.MolMR, return_dtype=pl.Float32).alias('MolMR'),
-        pl.col('Mol').map_elements(Descriptors.TPSA, return_dtype=pl.Float32).alias('TPSA'),
+        pl.col('Mol').map_elements(Descriptors.MolWt, return_dtype=pl.Float64).alias('MolWt'),
+        pl.col('Mol').map_elements(Descriptors.MolLogP, return_dtype=pl.Float64).alias('MolLogP'),
+        pl.col('Mol').map_elements(Descriptors.MolMR, return_dtype=pl.Float64).alias('MolMR'),
+        pl.col('Mol').map_elements(Descriptors.TPSA, return_dtype=pl.Float64).alias('TPSA'),
         pl.col('Mol').map_elements(Descriptors.HeavyAtomCount, return_dtype=pl.Int16).alias('NumHeavyAtom'),
         pl.col('Mol').map_elements(Descriptors.NumHAcceptors, return_dtype=pl.Int16).alias('NumHAcceptors'),
         pl.col('Mol').map_elements(Descriptors.NumHDonors, return_dtype=pl.Int16).alias('NumHDonors'),
